more testing of DropHandlerLogic
